+++ src/cli/cmd/iexec-storage.js
@@ -20,6 +20,7 @@ const {
   Spinner,
   getPropertyFormChain,
   getSmsUrlFromChain,
+  optionCreator,
 } = require('../utils/cli-helper');
 const { loadChain, connectKeystore } = require('../utils/chains');
 const { Keystore } = require('../utils/keystore');
@@ -33,6 +34,7 @@ initStorage
   .option(...option.chain())
   .option(...option.forceUpdateSecret())
   .option(...option.storageToken())
+  .addOption(optionCreator.teeFramework())
   .description(desc.initStorage())
   .action(async (provider, opts) => {
     await checkUpdate(opts);
@@ -45,7 +47,9 @@ initStorage
         keystore.accounts(),
       ]);
       const { contracts } = chain;
-      const smsURL = getSmsUrlFromChain(chain);
+      const smsURL = getSmsUrlFromChain(chain, {
+        teeFramework: opts.teeFramework,
+      });
       const resultProxyURL = getPropertyFormChain(chain, 'resultProxy');
 
       const providerName = provider || 'default';
@@ -105,6 +109,7 @@ addWalletLoadOptions(checkStorage);
 checkStorage
   .option(...option.chain())
   .option(...option.user())
+  .addOption(optionCreator.teeFramework())
   .description(desc.checkStorage())
   .action(async (provider, opts) => {
     await checkUpdate(opts);
@@ -117,7 +122,9 @@ checkStorage
         keystore.accounts(),
       ]);
       const { contracts } = chain;
-      const smsURL = getSmsUrlFromChain(chain);
+      const smsURL = getSmsUrlFromChain(chain, {
+        teeFramework: opts.teeFramework,
+      });
       const providerName = provider || 'default';
       const tokenKeyName = getStorageTokenKeyName(providerName);
       const userAddress = opts.user || address;